Have the secure credentials in separate header files.
